% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENQview_lite.R, R/ENQview_lite_with_module.R
\name{ENQview_lite}
\alias{ENQview_lite}
\title{ENQview_lite（モジュール版統合）— df 指定/アップロード入口を共通モジュールに置換}
\usage{
ENQview_lite(df = NULL)

ENQview_lite(df = NULL)
}
\description{
ENQview_lite（モジュール版統合）— df 指定/アップロード入口を共通モジュールに置換

ENQview_lite（モジュール版統合）— df 指定/アップロード入口を共通モジュールに置換
}
