% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_speMCA.R, R/Shiny_speMCA_with_module.R
\name{Shiny_speMCA}
\alias{Shiny_speMCA}
\title{speMCA用の変数選択app（モジュール版統合）}
\usage{
Shiny_speMCA(df = NULL)

Shiny_speMCA(df = NULL)
}
\description{
df が指定されていればそれを使用、未指定なら DataSourceModule 経由で .rda/.RData/.rds をアップロードして選択。
以降の本体処理は src$data() を唯一のデータソースとして参照します。

df が指定されていればそれを使用、未指定なら DataSourceModule 経由で .rda/.RData/.rds をアップロードして選択。
以降の本体処理は src$data() を唯一のデータソースとして参照します。
}
